refactor(keybinds): Change debouncer for shortcut to use_global_shortcut
